Fixed updating and loading of filtered stock
